requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.22
cassandra-driver==3.29.2
tiktoken==0.7.0
pypdf==4.3.1
requests==2.32.3
